% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{build_pcn}
\alias{build_pcn}
\title{Build a protein contact network from coordinates}
\usage{
build_pcn(coords, cutoff = 7, directed = FALSE)
}
\arguments{
\item{coords}{data.frame with columns: resno, resname, x, y, z}

\item{cutoff}{numeric distance threshold (Å) to define a contact}

\item{directed}{logical; whether to return a directed graph (default FALSE)}
}
\value{
list with elements: graph (igraph), adjacency (matrix), coords (data.frame)
}
\description{
Build a protein contact network from coordinates
}
