% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{coords_from_pdb}
\alias{coords_from_pdb}
\title{Extract residue coordinates from a bio3d pdb object}
\usage{
coords_from_pdb(pdb, chain = "A", atom = "CA")
}
\arguments{
\item{pdb}{a `bio3d` pdb object (from `bio3d::read.pdb`)}

\item{chain}{single-letter chain identifier (e.g., "A")}

\item{atom}{atom name to select (e.g., "CA", "CB")}
}
\value{
data.frame with resno, resname, x, y, z
}
\description{
Extract residue coordinates from a bio3d pdb object
}
