% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-batch.R
\name{pcn_batch}
\alias{pcn_batch}
\title{Batch processing of PDB jobs from a TSV}
\usage{
pcn_batch(tsv_file, download_path = tempdir(), progress = TRUE)
}
\arguments{
\item{tsv_file}{path to a tab-separated file}

\item{download_path}{directory for downloaded PDBs (default tempdir())}

\item{progress}{logical; print progress messages}
}
\value{
named list of PCN objects
}
\description{
Read a tab-delimited file with columns: `PDB`, `CHAIN`, `ATOM`, `CUTOFF` and
run PCN construction for each row. Returns a list of results.
}
