% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-analysis.R
\name{pcn_communities}
\alias{pcn_communities}
\title{Detect residue communities (modules) in a PCN}
\usage{
pcn_communities(pcn, method = c("louvain", "fast_greedy", "walktrap"))
}
\arguments{
\item{pcn}{A PCN object}

\item{method}{One of "louvain","fast_greedy","walktrap".}
}
\description{
Detect residue communities (modules) in a PCN
}
