% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-analysis.R
\name{pcn_compare}
\alias{pcn_compare}
\title{Compare two PCNs (edge-level)}
\usage{
pcn_compare(pcn1, pcn2)
}
\arguments{
\item{pcn1}{First PCN}

\item{pcn2}{Second PCN}
}
\description{
Compare two PCNs (edge-level)
}
