% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{pcn_edges}
\alias{pcn_edges}
\title{Get edge list (data.frame) from a PCN object}
\usage{
pcn_edges(pcn)
}
\arguments{
\item{pcn}{object returned by `build_pcn()`}
}
\value{
data.frame with columns: from, to (undirected)
}
\description{
Get edge list (data.frame) from a PCN object
}
