% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-analysis.R
\name{pcn_export_graphml}
\alias{pcn_export_graphml}
\title{Export a PCN graph to GraphML}
\usage{
pcn_export_graphml(pcn, file)
}
\arguments{
\item{pcn}{PCN object}

\item{file}{Output path ending with .graphml}
}
\description{
Export a PCN graph to GraphML
}
