% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{pcn_from_file}
\alias{pcn_from_file}
\title{Build a PCN from a local PDB file}
\usage{
pcn_from_file(file, chain = "A", atom = "CA", cutoff = 7)
}
\arguments{
\item{file}{path to a .pdb file}

\item{chain}{chain identifier}

\item{atom}{atom name (e.g., "CA")}

\item{cutoff}{numeric Å cutoff}
}
\description{
Build a PCN from a local PDB file
}
