% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{pcn_from_pdb}
\alias{pcn_from_pdb}
\title{Build a PCN from a PDB ID (download via bio3d)}
\usage{
pcn_from_pdb(pdb_id, chain = "A", atom = "CA", cutoff = 7, path = tempdir())
}
\arguments{
\item{pdb_id}{four-character PDB identifier}

\item{chain}{chain identifier}

\item{atom}{atom name (e.g., "CA")}

\item{cutoff}{numeric Å cutoff}

\item{path}{directory to store downloaded PDB}
}
\value{
list as in `build_pcn()`
}
\description{
Build a PCN from a PDB ID (download via bio3d)
}
