% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-core.R
\name{pcn_to_adjacency}
\alias{pcn_to_adjacency}
\title{Get adjacency matrix from a PCN object}
\usage{
pcn_to_adjacency(pcn)
}
\arguments{
\item{pcn}{object returned by `build_pcn()`}
}
\description{
Get adjacency matrix from a PCN object
}
