% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCNR-package.R
\docType{package}
\name{PCNR-package}
\alias{PCNR}
\alias{PCNR-package}
\title{PCNR: Protein Contact Networks in R}
\description{
Build residue-level protein contact networks (PCNs) from PDB IDs/files,
compute adjacency, analyze centralities/communities, and export graphs.
}
\author{
\strong{Maintainer}: NIRJHAR BHATTACHARYYA \email{nirjhar.suvo.97@gmail.com}

Other contributors:
\itemize{
  \item PCNR Contributors [contributor]
}

}
\keyword{internal}
