% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-plot.R
\name{plot_pcn}
\alias{plot_pcn}
\title{Quick plot for a PCN igraph}
\usage{
plot_pcn(g, layout = "kk")
}
\arguments{
\item{g}{igraph object (e.g., pcn$graph)}

\item{layout}{character: one of 'kk', 'fr', or 'circle'}
}
\description{
Quick plot for a PCN igraph
}
