% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-io.R
\name{write_pcn_adjacency}
\alias{write_pcn_adjacency}
\title{Write adjacency matrix to CSV}
\usage{
write_pcn_adjacency(pcn, file)
}
\arguments{
\item{pcn}{PCN object}

\item{file}{output path}
}
\description{
Write adjacency matrix to CSV
}
