% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcn-io.R
\name{write_pcn_edges}
\alias{write_pcn_edges}
\title{Write edge list to CSV}
\usage{
write_pcn_edges(pcn, file)
}
\arguments{
\item{pcn}{PCN object}

\item{file}{output path}
}
\description{
Write edge list to CSV
}
